Splitting up components.
