+++ graylog2-web-interface/src/components/events/bulk-replay/types.ts
@@ -1,0 +1,4 @@
+export type Event = {
+  id: string,
+  message: string,
+}
